Dashboard styling complete pending Highcharts import
