+++ client/src/Components/Header/Navbar.css
@@ -85,7 +85,7 @@
   display: none;
 }
 
-@media screen and (max-width: 960px) {
+@media screen and (max-width: 770px) {
   .NavbarItems {
     position: relative;
   }
